Contract updated

+++ src/AAE_Token.sol
@@ -4,14 +4,13 @@ pragma solidity 0.8.23;
 /// @title AAE_Token: An ERC20 token that is Ownable, Pausable, and Burnable
 /// @dev This contract implements an ERC20 token with extensions for pausability, burnability, and ownership transfer.
 contract AAE_Token {
-
     string public constant name = "AAE_Token";
     string public constant symbol = "AAET";
     uint8 public constant decimals = 18;
     /// @notice Total supply of tokens, initially set to 1,000,000 tokens (including decimals)
-    uint256 public totalSupply = 1000000 * 10**uint256(decimals);
+    uint256 public totalSupply = 1000000 * 10 ** decimals;
     /// @notice Initial circulating supply set to 100,000 tokens (including decimals)
-    uint256 public initialSupply = 100000 * 10**uint256(decimals);
+    uint256 public initialSupply = 100000 * 10 ** decimals;
 
     /// @notice Address of the contract owner
     address public owner;
@@ -26,9 +25,16 @@ contract AAE_Token {
     /// @notice Emitted when tokens are transferred, including zero value transfers
     event Transfer(address indexed from, address indexed to, uint256 value);
     /// @notice Emitted when a successful approval of allowances is made
-    event Approval(address indexed owner, address indexed spender, uint256 value);
+    event Approval(
+        address indexed owner,
+        address indexed spender,
+        uint256 value
+    );
     /// @notice Emitted when ownership of the contract is transferred
-    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
+    event OwnershipTransferred(
+        address indexed previousOwner,
+        address indexed newOwner
+    );
     /// @notice Emitted when the contract is paused
     event Paused();
     /// @notice Emitted when the contract is unpaused
@@ -80,21 +86,23 @@ contract AAE_Token {
     /// @notice Burns a specific amount of tokens from the caller's balance
     /// @dev Reduces the total supply of tokens
     /// @param amount The amount of tokens to be burned
-        function burn(uint256 amount) public whenNotPaused {
-            require(amount <= balances[msg.sender], "Insufficient balance");
-            // Adjust this line to modify totalSupply instead of initialSupply
-            totalSupply -= amount;
-            balances[msg.sender] -= amount;
-            emit Burn(msg.sender, amount);
-        }
-
+    function burn(uint256 amount) public whenNotPaused {
+        require(amount <= balances[msg.sender], "Insufficient balance");
+        // Adjust this line to modify totalSupply instead of initialSupply
+        totalSupply -= amount;
+        balances[msg.sender] -= amount;
+        emit Burn(msg.sender, amount);
+    }
 
     /// @notice Transfers a specific amount of tokens to a specified address
     /// @dev Requires the recipient address to be non-zero and the sender to have a sufficient balance
     /// @param to The recipient address
     /// @param amount The amount of tokens to transfer
     /// @return A boolean value indicating success
-    function transfer(address to, uint256 amount) public whenNotPaused returns (bool) {
+    function transfer(
+        address to,
+        uint256 amount
+    ) public whenNotPaused returns (bool) {
         require(to != address(0), "Transfer to the zero address");
         require(balances[msg.sender] >= amount, "Insufficient balance");
 
@@ -124,7 +132,10 @@ contract AAE_Token {
     ) public whenNotPaused returns (bool) {
         require(to != address(0), "Transfer to the zero address");
         require(balances[from] >= amount, "Insufficient balance");
-        require(allowed[from][msg.sender] >= amount, "Transfer amount exceeds allowance");
+        require(
+            allowed[from][msg.sender] >= amount,
+            "Transfer amount exceeds allowance"
+        );
 
         balances[from] -= amount;
         balances[to] += amount;
@@ -138,7 +149,10 @@ contract AAE_Token {
     /// @param spender The address which will spend the funds
     /// @param amount The amount of tokens to be spent
     /// @return A boolean value indicating success
-    function approve(address spender, uint256 amount) public whenNotPaused returns (bool) {
+    function approve(
+        address spender,
+        uint256 amount
+    ) public whenNotPaused returns (bool) {
         allowed[msg.sender][spender] = amount;
         emit Approval(msg.sender, spender, amount);
         return true;
@@ -148,7 +162,10 @@ contract AAE_Token {
     /// @param _owner The address of the token owner
     /// @param _spender The address which will spend the tokens
     /// @return The number of tokens still available for the spender
-    function allowance(address _owner, address _spender) public view returns (uint256) {
+    function allowance(
+        address _owner,
+        address _spender
+    ) public view returns (uint256) {
         return allowed[_owner][_spender];
     }
 }
